bibliotheque/requirements.txt
Django>=5.0,<6.0
psycopg2-binary
djangorestframework
django-cors-headers